+++ stable_baselines3/dqn/dqn.py
@@ -5,6 +5,7 @@ import torch as th
 from torch.nn import functional as F
 
 from stable_baselines3.common import logger
+from stable_baselines3.common.buffers import ReplayBuffer
 from stable_baselines3.common.off_policy_algorithm import OffPolicyAlgorithm
 from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback
 from stable_baselines3.common.utils import get_linear_fn, polyak_update
@@ -69,6 +70,8 @@ class DQN(OffPolicyAlgorithm):
         gradient_steps: int = 1,
         n_episodes_rollout: int = -1,
         optimize_memory_usage: bool = False,
+        replay_buffer_class: Optional[Type[ReplayBuffer]] = None,
+        replay_buffer_kwargs: Optional[Dict[str, Any]] = None,
         target_update_interval: int = 10000,
         exploration_fraction: float = 0.1,
         exploration_initial_eps: float = 1.0,
@@ -88,6 +91,8 @@ class DQN(OffPolicyAlgorithm):
             env,
             DQNPolicy,
             learning_rate,
+            replay_buffer_class,
+            replay_buffer_kwargs,
             buffer_size,
             learning_starts,
             batch_size,