+++ stable_baselines3/td3/td3.py
@@ -4,6 +4,7 @@ import torch as th
 from torch.nn import functional as F
 
 from stable_baselines3.common import logger
+from stable_baselines3.common.buffers import ReplayBuffer
 from stable_baselines3.common.noise import ActionNoise
 from stable_baselines3.common.off_policy_algorithm import OffPolicyAlgorithm
 from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback
@@ -72,6 +73,8 @@ class TD3(OffPolicyAlgorithm):
         n_episodes_rollout: int = 1,
         action_noise: Optional[ActionNoise] = None,
         optimize_memory_usage: bool = False,
+        replay_buffer_class: Optional[Type[ReplayBuffer]] = None,
+        replay_buffer_kwargs: Optional[Dict[str, Any]] = None,
         policy_delay: int = 2,
         target_policy_noise: float = 0.2,
         target_noise_clip: float = 0.5,
@@ -89,6 +92,8 @@ class TD3(OffPolicyAlgorithm):
             env,
             TD3Policy,
             learning_rate,
+            replay_buffer_class,
+            replay_buffer_kwargs,
             buffer_size,
             learning_starts,
             batch_size,